Exit immediately on non-zero status from pipelines
Prior to this change, the script allowed for silent failures of pipelines.
This change uses the set -e command to alter that default behavior.

This change also ignores a non-zero status when unmount is called to unmount
the USB device's first partition.  Testing showed that the umount command
returned a non-zero status on every execution because the partition was not
mounted.  Prior to this change, the same non-zero status was being ignored.

Related: https://github.com/calmh/smartos-platform-upgrade/issues/14